DRY file loading
